Fix #6137, dialog('open') causes form elements to reset on IE7.
Looking through the history, the behavior of appending to the body was
present in the initial commit without explanation. The only substantive
change thereafter was adding the next().length check in fbdb9f08,
fixing #3012.

uiDialog is already appended to the body in _create, and doing it again
in open seems only to cause problems. Removing it does not cause any
test regressions.