Reformat minDate/maxDate when dateFormat changes
The minDate/maxDate properties are assumed to be in the same date
format as the dateFormat of the datepicker.  If the datepicker is
later updated to use a new dateFormat, minDate/maxDate remain
in the old format and are not used properly by the datepicker.

With this change, when a new dateFormat is specified, minDate/maxDate
are reformatted with the new date format.  This allows the datepicker
to parse those dates correctly and properly restrict the range of the
datepicker.