Allowing theming of the buttons in dialogs. Also,
Sometimes you need the buttons inside of dialogs to be a little
different, like a cancel button having red text. The easiest way to
accomplish this is to allow passing in a class that can flow directly
to the button. This is easily accomplished. Also, using on/off with
name scoping is useful in case of more advanced theming.