Tabs: add support for <base> tag handling
Prevents lots of grief for developers if <base> tag is used and e.g. you want to load local tabs in a modal dialog (where the page containing the tabs is loaded via AJAX)

Basically just reads <base> tag HREF if it exists and compares it to anchor in isLocal. Can probably be optimized further.