Slider: Avoid the slider to be locked when both values are max (if range option is true). Fixed #9046 - Setting both values of range slider to maximum value, locks the slider
Basically, by the time both handles are set to the maximum value, I give
a focus on the first handle. Doing that, the lock is avoided because the
user will have control on it (first handle), allowing drag it to the
left whenever necessary. I also added a visual test.