refac: mv _initializeSubMenus into menubarMenuItem
More complicated refactor. We move the
initialization routine into the menubarMenuItem,
but then for the application of the eventing we
throw the context *back* into the menubar.

We do this so that the `this.options`
automagically works and, from an aesthetic
judgment, the concern about what happens between
submenus ( move to another menuItem's submenu or
to another menuItem ) is a concern of the menubar.