Resetting a flag that is used to know whether mouse clicks were processed.
If there are several autocomplete fields on a page, it is possible to select one of the results with a mouse click only on the first autocomplete. The subsequent ones do not respond to a mouse click, as we found out, because mouseHandled flag didn't get reset (as we found out, the autocomplete dropdown is a jQueryUI menu).