Fized resizable bug "Custom handles do not work"
I needed to use as handle a div outside the resizable element. So I tried to specify the jquery element as the documentation said but it did not work.
handles: { "s": $("#customResizableHandle") },
The documentation said in handles option:
Object:
The following keys are supported: { n, e, s, w, ne, se, sw, nw }. The value of any specified should be a jQuery selector matching the child element of the resizable to use as that handle. If the handle is not a child of the resizable, you can pass in the DOMElement or a valid jQuery object directly.

Note: When generating your own handles, each handle must have the ui-resizable-handle class, as well as the appropriate ui-resizable-{direction} class, .e.g., ui-resizable-s.

I also found the next ticket: 
Ticket: http://bugs.jqueryui.com/ticket/4310 (2009 bug, still in last version)

My edition fix that bug (but I need to test it thoroughly). Now I can use a handle outside the element resizable. I hope this help.
My solution is based in a previous solution from ylebre http://stackoverflow.com/questions/958419/custom-resizable-handles-in-jquery-ui
I had to make some modifications and adaptions to the current version.