Tooltip: on close and destroy only set title if empty or undefined. Fixed #8925 - Changes to title attribute are reverted on close
Ticket #8925 states that changes to the title attribute while the
tooltip is open are lost on tooltip close.

In the close and destroy functions, the title attribute is always
written if a value was stored in the element on open. It is possible
the attribute has changed and restoring the initial value may overwrite
the current value.

If the value is empty or undefined as set in open, do not set the title
attribute.

This fix has the limitation that if the user removed the title
attribute or set the value to an empty string the original title value
would be restored on close and destroy. Test cases 3 and 4 demonstrate
this limitation.