Autocomplete: Ignore ENTER and TAB events for items with 'ui-state-disabled'. Fixes #9695 - autocomplete: should not allow to select a disabled item of its menu with the ENTER key
When ‘ui-state-disabled’ has been added to the <li> elements of the
menu, ignore ENTER and TAB keydown events.

For TAB, also call event.preventDefault() so that the menu is not
dismissed when the user presses TAB on a disabled item.  If this is not
called, while select() is not called, the item will appear to have been
selected.