Droppable: Unset visible on disabled/unaccepting droppables
Droppables initially start off with visible undefined, and previously
could get in a state where visible was set to true while not accepting
a draggable. In this state, drag code will continue to update onover
and onout on the droppable instead of ignoring the droppable.

This is problematic if, for example, the droppable then starts
accepting the draggable: if the draggable is already over the droppable
then isover will already be set to true, and so the the droppable will
not get an over event.

Set visible back to the initial state of undefined if the droppable is
disabled or not accepting of the draggable.

Fixes #9758