Fix problems with content async ajax request
There were 2 problems, first af all, the check for "ui-tooltip-open" was done -before- the _delay call. This means that when the _open function was actually called, the user could have already moved his mouse out.
I've moved the check to the start of the _open function because I believe that's where it should actually always be checked.
Second, the mouseout events were only set in the _open function. This way, with the async ajax request, the mouseout events were never called when the tooltip wasn't actually opened (the user moved the mouse away before the content was able to update). I've set the events in the open function now as well.