Update datepicker.js
Add new setting showSelectorAtPos to control the calendar on which the month and year selectors are shown (works the same as showCurrentAtPos). Modifies the detection of "secondary" when calling _generateMonthYearHeader. Tested against both single row and multi row/column configurations. Enables the selection boxes to be in the same month as the one they are setting where showCurrentAtPos is non-zero. showSelectorAtPos is zero based to match showCurrentAtPos.