autocomplete: fix for IE11/Edge scrolling issues
IE11 and Edge will not prevent focus when you `preventDefault` on a mousedown
events; this was causing some undesireable behavior:

1. Depending on how long you clicked, the dropdown could close (IE11)

2. Scrolling down by clicking the scrollbar's down arrow would misbehave. The
list would pop back up to the top with the first item selected. (IE11, Edge)

Both these problems are fixed by making the dropdown un-focusable (by setting
the IE/Edge-only `unselectable` attribute to `on`).

I couldn't figure out a way to write a unit test for this, since it's IE only
and seems to require user interaction. You can verify the previous behavior
(and the fix) on demos/autocomplete/maxheight.html

Fixes #9638