Fix autocomplete normalize, preserve empty value
In the normalize method, an item using a data object (label/value) with a value set to an empty string (useful, e.g., for an error response or an item in the list that should be disabled), the value is erroneously set to use the label string, because "item.value" is evaluated as false. Rewriting the normalize condition to check whether "item.value" is either truthy or an empty string.