Effects: Fix a timing issue in a variable declaration
A jQuery object was created, chained & assigned to a variable that was then
accessed in a callback used inside of this chained definition. Due to a timing
difference in when the callback fired for the first time in latest jQuery
master, it was being called before the variable was defined.

This has been resolved now.